Add callback to reset filenode to table access method
This commit adds the callback `relation_reset_filenode` to
`TableAccessMethod` that is called when the table about to be dropped.

To support transctions, the table access method should implement this
callback to register the table for deletion, but not physically delete
the table, and register a transaction handler with `RegisterXactHandler`
that will remove the table completely on commit, and restore it on
rollback.