Add callback for unlinking file node of table
When dropping or truncating a table, the old file locator is unlinked
and a new file locator created for the relation. This is, however, done
without notifying the table access method that the unlinking is taking
place. This becomes a problem if the table access method has an
internal resource associated with the file locator that also needs to
be unlinked or cleaned up.

This commit add a callback to the table access method for resetting the
file locator when it is dropped and call it to unlink the file locator
so that the table access method and unlink the resource internally, if
necessary. The callback can be undefined (that is, NULL) if the table
access method do not require notification that a resource is unlinked.