Use pstrdup instead of strdup in postmaster
There is no check for the return value of strdup() in
BackendInitialize() which means that it is possible to get a
segmentation fault when starting a backend if the system is low on
memory. Instead, use pstrdup() for the allocations that are within the
same process since this cannot return NULL and instead fail with an
error, aborting the start of a backend.