Use union as Datum for type-safety
Datum is currently an intptr_t to be able to represent both pointers to
actual values and references to values. However, if a Datum is used
"bare" (without using the DatumGetXYZ functions), it can accidentally
invoke an implicit integer conversion and treat a pointer as a value.

This is normally not a big problem since accidentally treating, e.g.,
an int32 as a Datum is not going to break anything, but if it is
missing when using datum stored by reference it is going to create
problems.

This commit changes this by turning the Datum into a union and passing
it by value. Modern compilers will be able to pass this in a register
and will generate the same code for both constructions, but accidental
omissions of DatumGetXYZ will be caught at compile-time.